Refactored files to make things clearer and more manageable
