+++ src/main.py
@@ -1,11 +1,7 @@
 import tkinter as tk
 import ttkbootstrap as ttk
 from tkinter import filedialog
-from astropy.io import fits
-from astropy.visualization import LogStretch
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
-from matplotlib.figure import Figure
-import matplotlib.pyplot as plt
+from src.widgets import image_controller as ic
 
 
 # Create Main Tkinter Window
@@ -16,21 +12,22 @@ class MainWindow(tk.Tk):
         self.geometry("250x250")
 
         self.menu_controller = MenuBar(self)
-        self.image_controller = ImageController(self)
+        self.image_controller = ic.ImageController(self)
 
         self.config(menu=self.menu_controller.menu)
 
     # Main run function for app
     # Place all functions of the app here
     def run(self):
-        if self.menu_controller.file_name != "":
+        if self.menu_controller.file_opened and self.menu_controller.file_name != "":
             file_name = self.menu_controller.file_name
             self.image_controller.open_image(file_name)
             self.menu_controller.file_name = ""
+            self.menu_controller.file_opened = False
 
         # Do not remove or edit, required for run loop to function
         self.update()
-        self.after(0, self.run)
+        self.after(100, self.run)
 
 
 # Create Menu bar for tkinter window
@@ -62,67 +59,6 @@ class MenuBar(ttk.Frame):
             title="Select .fits file",
             filetypes=(("Fits files", "*.fits"), ("All files", "*.*")),
         )
-
-
-# Create Image Controller Frame
-class ImageController(ttk.Frame):
-    gridX = 0
-    gridY = 0
-
-    open_images = []
-
-    def __init__(self, parent):
-        ttk.Frame.__init__(self, parent)
-        self.grid(column=0, row=1)
-        test_label = ttk.Button(self, text="Button 1", bootstyle="success")
-        test_label.grid(column=2, row=0)
-
-    # Open image file based on path selected
-    def open_image(self, file_path):
-        new_image = ImageFrame(self, file_path, self.gridX, self.gridY)
-        self.open_images.append(new_image)
-
-        if self.gridX > self.gridY:
-            self.gridY += 1
-            self.gridX = 0
-        else:
-            self.gridX += 1
-
-
-# Create an Image Frame
-class ImageFrame(ttk.Frame):
-    image_file = None
-
-    def __init__(self, parent, file_path, x, y):
-        ttk.Frame.__init__(self, parent)
-        self.parent = parent
-        self.image_file = file_path
-        self.grid(column=x, row=y)
-        self.handle_image()
-
-    def handle_image(self):
-        # Read the .fits file
-        hdu_list = fits.open(self.image_file)
-        image_data = hdu_list[0].data
-
-        # Apply logarithmic scaling to the image data
-        log_stretch = LogStretch()
-        scaled_data = log_stretch(image_data)
-
-        # Create a matplotlib figure
-        fig = Figure(figsize=(5, 5), dpi=100)
-        ax = fig.add_subplot(111)
-
-        # Render the scaled image data onto the figure
-        cax = ax.imshow(scaled_data, cmap="gray", origin="lower")
-
-        # Add a colourbar
-        fig.colorbar(cax)
-
-        # Embed the matplotlib figure in the tkinter window
-        canvas = FigureCanvasTkAgg(fig, master=self)
-        canvas.draw()
-        canvas.get_tk_widget().grid(column=0, row=0)
 
 
 # Run the main app
